fix: close modal when back button in tablet is touched
